Modified definations of PEStruct. WORD Characteristics --> DWORD Characteristics
Abandon old method (scmpi(secname,".text"))
Use new method (sec.Charasteristics & 0x20000000)

Which fixs some delphi problem.

+++ src/WProtect/PE/PESection.cpp
@@ -184,7 +184,7 @@ BYTE* CPESection::GetSectionData(int nIndex, DWORD *rdwSize)
 	return lpBuffer;
 }
 
-WORD CPESection::GetCharacteristics(int nIndex)
+DWORD CPESection::GetCharacteristics(int nIndex)
 {
     return GetSection(nIndex)->Characteristics;
 }
